Add default template and handle no file
