+++ lib/read.js
@@ -13,7 +13,7 @@ var getTemplateFile = function getTemplateFile(template) {
       return "templates/css.js";
 
     default:
-      return "";
+      return "templates/default.js";
   }
 };
